chore(core): move test project generator to its own file
Move test project generator and its associated unit tests to separate files
