+++ packages/core/src/generators/test/generator.ts
@@ -2,7 +2,7 @@ import { Tree } from '@nrwl/devkit';
 
 import { DotNetClient, dotnetFactory } from '@nx-dotnet/dotnet';
 
-import { GenerateTestProject } from '../utils/generate-project';
+import { GenerateTestProject } from '../utils/generate-test-project';
 import { NxDotnetGeneratorSchema } from './schema';
 
 export default function (
